+++ truffle/src/com.oracle.truffle.api.interop.java.test/src/com/oracle/truffle/api/interop/java/test/JavaInteropNullTest.java
@@ -54,7 +54,7 @@ public class JavaInteropNullTest {
     class TestRootNode extends RootNode {
 
         private final TruffleObject receiver;
-        @Child Node execute = Message.createExecute(1).createNode();
+        @Child Node execute = Message.EXECUTE.createNode();
 
         TestRootNode(TruffleObject receiver) {
             super(null);