+++ truffle/src/com.oracle.truffle.api.interop/src/com/oracle/truffle/api/interop/ForeignAccess.java
@@ -972,9 +972,9 @@ public final class ForeignAccess {
         }
 
         /**
-         * Handles {@link Message#createExecute(int)} messages.
+         * Handles {@link Message#EXECUTE} messages.
          *
-         * @param argumentsLength number of parameters the messages has been created for
+         * @param argumentsLength do not use, always 0
          * @return call target to handle the message or <code>null</code> if this message is not
          *         supported
          * @since 0.30
@@ -984,9 +984,9 @@ public final class ForeignAccess {
         }
 
         /**
-         * Handles {@link Message#createInvoke(int)} messages.
+         * Handles {@link Message#INVOKE} messages.
          *
-         * @param argumentsLength number of parameters the messages has been created for
+         * @param argumentsLength do not use, always 0
          * @return call target to handle the message or <code>null</code> if this message is not
          *         supported
          * @since 0.30
@@ -996,9 +996,9 @@ public final class ForeignAccess {
         }
 
         /**
-         * Handles {@link Message#createNew(int)} messages.
+         * Handles {@link Message#NEW} messages.
          *
-         * @param argumentsLength number of parameters the messages has been created for
+         * @param argumentsLength do not use, always 0
          * @return call target to handle the message or <code>null</code> if this message is not
          *         supported
          * @since 0.30
@@ -1173,9 +1173,9 @@ public final class ForeignAccess {
         CallTarget accessWrite();
 
         /**
-         * Handles {@link Message#createExecute(int)} messages.
+         * Handles {@link Message#EXECUTE} messages.
          *
-         * @param argumentsLength number of parameters the messages has been created for
+         * @param argumentsLength do not use, always 0
          * @return call target to handle the message or <code>null</code> if this message is not
          *         supported
          * @since 0.26
@@ -1183,9 +1183,9 @@ public final class ForeignAccess {
         CallTarget accessExecute(int argumentsLength);
 
         /**
-         * Handles {@link Message#createInvoke(int)} messages.
+         * Handles {@link Message#INVOKE} messages.
          *
-         * @param argumentsLength number of parameters the messages has been created for
+         * @param argumentsLength do not use, always 0
          * @return call target to handle the message or <code>null</code> if this message is not
          *         supported
          * @since 0.26
@@ -1193,9 +1193,9 @@ public final class ForeignAccess {
         CallTarget accessInvoke(int argumentsLength);
 
         /**
-         * Handles {@link Message#createNew(int)} messages.
+         * Handles {@link Message#NEW} messages.
          *
-         * @param argumentsLength number of parameters the messages has been created for
+         * @param argumentsLength do not use, always 0
          * @return call target to handle the message or <code>null</code> if this message is not
          *         supported
          * @since 0.26
@@ -1293,12 +1293,12 @@ public final class ForeignAccess {
         private static CallTarget accessMessage(StandardFactory factory, Message msg) {
             if (msg instanceof KnownMessage) {
                 switch (msg.hashCode()) {
-                    case Execute.EXECUTE:
-                        return factory.accessExecute(((Execute) msg).getArity());
-                    case Execute.INVOKE:
-                        return factory.accessInvoke(((Execute) msg).getArity());
-                    case Execute.NEW:
-                        return factory.accessNew(((Execute) msg).getArity());
+                    case Execute.HASH:
+                        return factory.accessExecute(0);
+                    case Invoke.HASH:
+                        return factory.accessInvoke(0);
+                    case New.HASH:
+                        return factory.accessNew(0);
                     case GetSize.HASH:
                         return factory.accessGetSize();
                     case HasKeys.HASH:
@@ -1362,12 +1362,12 @@ public final class ForeignAccess {
         private static CallTarget accessMessage(Factory26 factory, Message msg) {
             if (msg instanceof KnownMessage) {
                 switch (msg.hashCode()) {
-                    case Execute.EXECUTE:
-                        return factory.accessExecute(((Execute) msg).getArity());
-                    case Execute.INVOKE:
-                        return factory.accessInvoke(((Execute) msg).getArity());
-                    case Execute.NEW:
-                        return factory.accessNew(((Execute) msg).getArity());
+                    case Execute.HASH:
+                        return factory.accessExecute(0);
+                    case Invoke.HASH:
+                        return factory.accessInvoke(0);
+                    case New.HASH:
+                        return factory.accessNew(0);
                     case GetSize.HASH:
                         return factory.accessGetSize();
                     case HasSize.HASH: