+++ truffle/src/com.oracle.truffle.nfi.test/src/com/oracle/truffle/nfi/test/PointerNFITest.java
@@ -92,7 +92,7 @@ public class PointerNFITest extends NFITest {
     public static class FoldVector extends SendExecuteNode {
 
         public FoldVector() {
-            super("foldVector", "(POINTER, UINT32) : DOUBLE", 2);
+            super("foldVector", "(POINTER, UINT32) : DOUBLE");
         }
     }
 
@@ -114,7 +114,7 @@ public class PointerNFITest extends NFITest {
     public static class IncVector extends SendExecuteNode {
 
         public IncVector() {
-            super("incVector", "(POINTER, SINT32, DOUBLE) : VOID", 3);
+            super("incVector", "(POINTER, SINT32, DOUBLE) : VOID");
         }
     }
 
@@ -160,7 +160,7 @@ public class PointerNFITest extends NFITest {
 
     public static class TestSlowPath extends NFITestRootNode {
 
-        @Child Node execute = Message.createExecute(1).createNode();
+        @Child Node execute = Message.EXECUTE.createNode();
 
         @Child Node isPointer = Message.IS_POINTER.createNode();
         @Child Node isBoxed = Message.IS_BOXED.createNode();